v3: Fix InMemoryCache Clear() and LoadMany() concurrency bugs
